+++ src/processors/tone/AmpIRs.h
@@ -21,7 +21,7 @@ public:
     void prepare (double sampleRate, int samplesPerBlock) override;
     void processAudio (AudioBuffer<float>& buffer) override;
 
-    void getCustomComponents (OwnedArray<Component>& customComps) override;
+    bool getCustomComponents (OwnedArray<Component>& customComps) override;
 
     std::unique_ptr<XmlElement> toXML() override;
     void fromXML (XmlElement* xml, bool loadPosition) override;