+++ src/processors/utility/InputProcessor.cpp
@@ -32,7 +32,8 @@ void InputProcessor::processAudio (AudioBuffer<float>& buffer)
     rmsLevels[1] = buffer.getMagnitude (1 % numChannels, 0, numSamples);
 }
 
-void InputProcessor::getCustomComponents (OwnedArray<Component>& customComps)
+bool InputProcessor::getCustomComponents (OwnedArray<Component>& customComps)
 {
     customComps.add (std::make_unique<LevelMeterComponent> (rmsLevels));
+    return false;
 }